+++ uneval/__init__.py
@@ -1,4 +1,4 @@
-from .dsl import and_, or_, not_, in_, if_, for_, lambda_, λ_, fstr, fmt, var, expr, ExprType, lit
+from .dsl import and_, or_, not_, in_, if_, for_, lambda_, λ_, fstr, fmt, var, expr, ExprType, lit, scoped
 from .astbuild import to_ast
 from .lambdas import F, λ
 from .evaluation import evaluate, compiled
@@ -6,6 +6,7 @@ from .expression import Expression
 
 __all__ = [
     "expr",
+    "scoped",
     "evaluate",
     "compiled",
     "Expression",